Implement Main_KAnonymity class to do unit testing
